+++ packages/driver/cypress/e2e/commands/navigation.cy.js
@@ -97,15 +97,13 @@ describe('src/cy/commands/navigation', () => {
     })
 
     it('removes listeners', () => {
-      cy.log(Cypress.browser)
-      const unloadEvent = Cypress.browser.family === 'chromium' ? 'pagehide' : 'unload'
       const win = cy.state('window')
 
       const rel = cy.stub(win, 'removeEventListener')
 
       cy.reload().then(() => {
         expect(rel).to.be.calledWith('beforeunload')
-        expect(rel).to.be.calledWith(unloadEvent)
+        expect(rel).to.be.calledWith('unload')
       })
     })
 
@@ -415,10 +413,8 @@ describe('src/cy/commands/navigation', () => {
         const rel = cy.stub(win, 'removeEventListener')
 
         cy.go('back').then(() => {
-          const unloadEvent = cy.browser.family === 'chromium' ? 'pagehide' : 'unload'
-
           expect(rel).to.be.calledWith('beforeunload')
-          expect(rel).to.be.calledWith(unloadEvent)
+          expect(rel).to.be.calledWith('unload')
         })
       })
     })